refactor: cache leverage bracket and position risk
